refactor(app): move app-shell chrome into layout/
Moves logo.tsx, header/ (nav-bar, color-scheme-toggle), demo-alert.tsx,
error-boundary.tsx and code-highlight-provider.tsx out of the flat
shared/ folder into a new layout/ folder, since they are only used by
app/layout.tsx and app/provider.tsx to build the app shell. Updates
import paths accordingly. No behavioral changes.
